Move adapter to own file
