+++ Tests/JS-A+/AllTests.swift
@@ -10,49 +10,6 @@ import PromiseKit
 import JavaScriptCore
 
 @available(iOS 10.0, *)
-let resolved: @convention(block) (JSValue) -> JSPromise = { value in
-    return JSPromise(promise: .value(value))
-}
-
-@available(iOS 10.0, *)
-let rejected: @convention(block) (JSValue) -> JSPromise = { reason in
-    let error = JSPromise.JSError(reason: reason)
-    let promise = Promise<JSValue>(error: error)
-    return JSPromise(promise: promise)
-}
-
-@available(iOS 10.0, *)
-let deferred: @convention(block) () -> JSValue = {
-    
-    let context = JSContext.current()
-    
-    guard let object = JSValue(object: NSDictionary(), in: context) else {
-        fatalError("Couldn't create object")
-    }
-    
-    let pendingPromise = Promise<JSValue>.pending()
-    let jsPromise = JSPromise(promise: pendingPromise.promise)
-    
-    // promise
-    object.setObject(jsPromise, forKeyedSubscript: "promise" as NSString)
-    
-    // resolve
-    let resolve: @convention(block) (JSValue) -> Void = { value in
-        pendingPromise.resolver.fulfill(value)
-    }
-    object.setObject(resolve, forKeyedSubscript: "resolve" as NSString)
-    
-    // reject
-    let reject: @convention(block) (JSValue) -> Void = { reason in
-        let error = JSPromise.JSError(reason: reason)
-        pendingPromise.resolver.reject(error)
-    }
-    object.setObject(reject, forKeyedSubscript: "reject" as NSString)
-    
-    return object
-}
-
-@available(iOS 10.0, *)
 class AllTests: XCTestCase {
     
     func testAll() {
@@ -90,9 +47,9 @@ class AllTests: XCTestCase {
         guard let adapter = JSValue(object: NSDictionary(), in: context) else {
             fatalError("Couldn't create adapter")
         }
-        adapter.setObject(resolved, forKeyedSubscript: "resolved" as NSString)
-        adapter.setObject(rejected, forKeyedSubscript: "rejected" as NSString)
-        adapter.setObject(deferred, forKeyedSubscript: "deferred" as NSString)
+        adapter.setObject(JSAdapter.resolved, forKeyedSubscript: "resolved" as NSString)
+        adapter.setObject(JSAdapter.rejected, forKeyedSubscript: "rejected" as NSString)
+        adapter.setObject(JSAdapter.deferred, forKeyedSubscript: "deferred" as NSString)
         
         // Evaluate contents of `build.js`, which exposes `runTests` in the global context
         context.evaluateScript(script)
